+++ apps/portal-web/src/pages/api/file/queryFileTransferProgress.ts
@@ -20,45 +20,45 @@ import { getClient } from "src/utils/client";
 import { route } from "src/utils/route";
 import { handlegRPCError } from "src/utils/server";
 
-export const TerminateFilesTransferSchema = typeboxRouteSchema({
-  method: "POST",
+export const TransferInfo = Type.Object({
+  toCluster: Type.String(),
+  filePath: Type.String(),
+  transferSizeKb: Type.Number(),
+  progress: Type.Number(),
+  speedKBps: Type.Number(),
+  remainingTimeSeconds: Type.Number(),
+});
 
-  body: Type.Object({
-    fromCluster: Type.String(),
-    toCluster: Type.String(),
-    fromPath: Type.String(),
+
+export const QueryFileTransferProgressSchema = typeboxRouteSchema({
+  method: "GET",
+
+  query: Type.Object({
+    cluster: Type.String(),
   }),
 
   responses: {
-    204: Type.Null(),
-    415: Type.Object({
-      code: Type.Literal("SCOW-SYNC-TERMINATE_CMD_FAILED"),
-      // stderr of the scow-sync command
-      error: Type.String(),
-    }),
+    200: Type.Object({ result: Type.Array(TransferInfo) }),
     400: Type.Object({ code: Type.Literal("INVALID_CLUSTER") }),
+    415: Type.Object({ code: Type.Literal("SCOW-SYNC-QUERY_CMD_FAILED") }),
   },
 });
 
 const auth = authenticate(() => true);
 
-export default route(TerminateFilesTransferSchema, async (req, res) => {
+export default route(QueryFileTransferProgressSchema, async (req, res) => {
 
   const info = await auth(req, res);
 
   if (!info) { return; }
 
-  const { fromCluster, toCluster, fromPath } = req.body;
+  const { cluster } = req.query;
 
   const client = getClient(FileServiceClient);
-
-  return asyncUnaryCall(client, "terminateFilesTransfer", {
-    userId: info.identityId,
-    fromCluster: fromCluster,
-    toCluster: toCluster,
-    fromPath: fromPath,
-  }).then(() => ({ 204: null }), handlegRPCError({
-    [status.INTERNAL]: (e) => ({ 415: { code: "SCOW-SYNC-TERMINATE_CMD_FAILED" as const, error: e.details } }),
+  return asyncUnaryCall(client, "queryFileTransfer", {
+    cluster, userId: info.identityId,
+  }).then((results) => ({ 200: { result: results.transferInfos } }), handlegRPCError({
     [status.NOT_FOUND]: () => ({ 400: { code: "INVALID_CLUSTER" as const } }),
+    [status.INTERNAL]: () => ({ 415: { code: "SCOW-SYNC-QUERY_CMD_FAILED" as const } }),
   }));
 });
