+++ protos/portal/file.proto
@@ -165,7 +165,7 @@ message ExistsResponse {
   bool exists = 1;
 }
 
-message StartFilesTransferRequest {
+message StartFileTransferRequest {
   string from_cluster = 1;
   string to_cluster = 2;
   string user_id = 3;
@@ -173,11 +173,11 @@ message StartFilesTransferRequest {
   string to_path = 5;
 }
 
-message StartFilesTransferResponse {
+message StartFileTransferResponse {
 
 }
 
-message QueryFilesTransferRequest {
+message QueryFileTransferRequest {
   string cluster = 1;
   string user_id = 2;
 }
@@ -191,18 +191,18 @@ message TransferInfo {
   uint64 remaining_time_seconds = 6;
 }
 
-message QueryFilesTransferResponse {
+message QueryFileTransferResponse {
   repeated TransferInfo transfer_infos = 1;
 }
 
-message TerminateFilesTransferRequest {
+message TerminateFileTransferRequest {
   string from_cluster = 1;
   string to_cluster = 2;
   string user_id = 3;
   string from_path = 4;
 }
 
-message TerminateFilesTransferResponse {
+message TerminateFileTransferResponse {
 
 }
 
@@ -236,11 +236,11 @@ service FileService {
 
   rpc Exists(ExistsRequest) returns (ExistsResponse);
 
-  rpc StartFilesTransfer(StartFilesTransferRequest) returns (StartFilesTransferResponse);
+  rpc StartFileTransfer(StartFileTransferRequest) returns (StartFileTransferResponse);
 
-  rpc QueryFilesTransfer(QueryFilesTransferRequest) returns (QueryFilesTransferResponse);
+  rpc QueryFileTransfer(QueryFileTransferRequest) returns (QueryFileTransferResponse);
 
-  rpc TerminateFilesTransfer(TerminateFilesTransferRequest) returns (TerminateFilesTransferResponse);
+  rpc TerminateFileTransfer(TerminateFileTransferRequest) returns (TerminateFileTransferResponse);
 
   rpc CheckTransferKey(CheckTransferKeyRequest) returns (CheckTransferKeyResponse);
 }
